Add abstract themes

+++ compile-css.sh
@@ -15,7 +15,11 @@ do
     --dest $static_dir/_data_uris.scss \
     --dest publik-base-theme/static/includes/_data_uris.scss
 
-  echo "Compiling CSS for $theme..."
-  cd $static_dir && sassc style.scss style.css
-	rm -rf $static_dir/.sass-cache/
+  if [[ "$theme" = __* ]]; then
+    echo "Skipping CSS compilation for abstract theme $theme"
+  else
+    echo "Compiling CSS for $theme..."
+    cd $static_dir && sassc style.scss style.css
+    rm -rf $static_dir/.sass-cache/
+  fi
 done
